Fix execution on arm hosts (fix #114 and #132)

+++ mount_image.sh
@@ -81,10 +81,14 @@ fi
 
 mv "${mount}/etc/resolv.conf" "${mount}/etc/_resolv.conf"
 cp /etc/resolv.conf "${mount}/etc/resolv.conf"
-cp /usr/bin/qemu-arm-static0 ${mount}/usr/bin/qemu-arm-static0
-cp /usr/bin/qemu-arm-static ${mount}/usr/bin/qemu-arm-static
-cp /usr/bin/qemu-aarch64-static0 ${mount}/usr/bin/qemu-aarch64-static0
-cp /usr/bin/qemu-aarch64-static ${mount}/usr/bin/qemu-aarch64-static
+if [ -e /usr/bin/qemu-arm-static0 ]; then
+    cp /usr/bin/qemu-arm-static0 ${mount}/usr/bin/qemu-arm-static0
+    cp /usr/bin/qemu-arm-static ${mount}/usr/bin/qemu-arm-static
+fi
+if [ -e /usr/bin/qemu-aarch64-static0 ]; then
+    cp /usr/bin/qemu-aarch64-static0 ${mount}/usr/bin/qemu-aarch64-static0
+    cp /usr/bin/qemu-aarch64-static ${mount}/usr/bin/qemu-aarch64-static
+fi
 if [ -e "${mount}/etc/ld.so.preload" ]; then
     cp "${mount}/etc/ld.so.preload" "${mount}/etc/_ld.so.preload"
     echo > "${mount}/etc/ld.so.preload"
